Testings murat

+++ src/main/java/io/swagger/controller/BankAccountController.java
@@ -162,7 +162,7 @@ public class BankAccountController {
     @CrossOrigin
     @RequestMapping(method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE,value="/getBankAccount/{IBAN}/Deposit")
     public ResponseEntity accountDeposit(@PathVariable("IBAN") String IBAN, @RequestBody Transaction transaction) {
-        TransactionInfo transactionInfo = bankAccountService.AccountDeposit(IBAN, transaction.getAmount());
+        TransactionInfo transactionInfo = bankAccountService.AccountDeposit(IBAN, transaction.getAmount().doubleValue());
 
         if(transactionInfo != null) {
             return new ResponseEntity<TransactionInfo>(transactionInfo,HttpStatus.ACCEPTED);
@@ -190,7 +190,7 @@ public class BankAccountController {
     @CrossOrigin
     @RequestMapping(method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE,value="/getBankAccount/{IBAN}/Withdraw")
     public ResponseEntity accountWithdraw(@PathVariable("IBAN") String IBAN, @RequestBody Transaction transaction) {
-        TransactionInfo transactionInfo = bankAccountService.AccountWithdraw(IBAN, transaction.getAmount());
+        TransactionInfo transactionInfo = bankAccountService.AccountWithdraw(IBAN, transaction.getAmount().doubleValue());
 
         if(transactionInfo != null) {
             return new ResponseEntity<TransactionInfo>(transactionInfo,HttpStatus.ACCEPTED);
